timeUsed is a float

+++ backend/src/graphql/model/GdtEntryList.ts
@@ -28,6 +28,6 @@ export class GdtEntryList {
   @Field(() => Float)
   gdtSum: number
 
-  @Field(() => Int)
+  @Field(() => Float)
   timeUsed: number
 }
